refactor: header. Move router-hoc to routes
